pyproject.toml
[tool.poetry]
name = "api-books"
version = "0.1.0"
description = ""
authors = ["Makar Bialiauski <bialiauski@protonmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
SQLAlchemy = "^1.4.41"
asyncpg = "^0.26.0"
python-dotenv = "^0.21.0"
greenlet = "^1.1.3"

[tool.poetry.scripts]
app = "run:start_app"

[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
